made changes to deposit

+++ src/main/java/C13Group2/BankingAPI/service/DepositService.java
@@ -1,6 +1,9 @@
 package C13Group2.BankingAPI.service;
 
+import C13Group2.BankingAPI.dto.CreateDepositDTO;
+import C13Group2.BankingAPI.dto.UpdateDepositDTO;
 import C13Group2.BankingAPI.enums.DepositStatus;
+import C13Group2.BankingAPI.enums.Medium;
 import C13Group2.BankingAPI.enums.TransactionType;
 import C13Group2.BankingAPI.exceptions.ResourceNotFoundException;
 import C13Group2.BankingAPI.model.Account;
@@ -12,6 +15,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -63,15 +68,17 @@ public class DepositService {
     public Deposit getDepositById(Long depositId){
         return depositRepository.findById(depositId).orElse(null);
     }
-    public Deposit createDeposit(Long accountId, String medium, Double amount, String description){
+    public Deposit createDeposit(Long accountId,CreateDepositDTO createDepositDTO){
         verifyIfAccountExists(accountId);
+        LocalDate currentDate = LocalDate.now();
         Deposit deposit = new Deposit();
         deposit.setType(TransactionType.DEPOSIT);
         deposit.setStatus(DepositStatus.COMPLETED);
-        deposit.setMedium(medium);
-        deposit.setDescription(description);
-
+        deposit.setMedium(Medium.valueOf(String.valueOf(Medium.BALANCE)));
+        deposit.setDescription(createDepositDTO.getDescription());
+        deposit.setAmount(createDepositDTO.getAmount());
         deposit.setAccount(accountRepository.findById(accountId).get());
+        deposit.setTransaction_date(currentDate);
         Account addAmount = accountRepository.findById(accountId).get();
         addAmount.setBalance(addAmount.getBalance() + deposit.getAmount());
         accountRepository.save(addAmount);
@@ -80,10 +87,20 @@ public class DepositService {
         return depositRepository.save(deposit);
 
     }
-    public Deposit updateDeposit(Deposit deposit){
-        return depositRepository.save(deposit);
+    public Deposit updateDeposit(Long depositId,String exceptionMessage, UpdateDepositDTO updateDepositDTO){
+        verifyDeposit(depositId,exceptionMessage);
+        Deposit changeDeposit = depositRepository.findById(depositId).orElse(null);
+        changeDeposit.setDescription(updateDepositDTO.getDescription());
+
+        return depositRepository.save(changeDeposit);
     }
-    public void deleteDeposit(Long depositId){
-        depositRepository.deleteById(depositId);
+    public Deposit deleteDeposit(Long depositId){
+        Deposit deposit = depositRepository.findById(depositId).orElse(null);
+        if (deposit != null) {
+            depositRepository.deleteById(depositId);
+            return deposit;
+        } else {
+            throw new EntityNotFoundException("Deposit with id " + depositId + " not found");
+        }
     }
 }
